중간 발표 전 프로젝트 업데이트3

+++ src/main/java/project/vinyl/service/ItemService.java
@@ -96,8 +96,11 @@ public class ItemService {
      **/
     //기능 : 삭제
     public void deleteItem(Long itemId, Long userId){
-        itemImgService.delete(itemId);
-        itemRepository.deleteById(itemId);
+        Item item = itemRepository.findById(itemId).orElseThrow(EntityExistsException::new);
+        if (item.getMember().getId()==userId){
+            itemImgService.delete(itemId);
+            itemRepository.deleteById(itemId);
+        }
     }
 
     /**
